order-processing-app/requirements.txt
python-dotenv==0.19.0
flask>=2.0.1
requests>=2.26.0
flask-sqlalchemy>=2.5.1
pyodbc>=4.0.31
pytz>=2021.1
apscheduler==3.7.0
scipy>=1.7.1
numpy>=1.12.1
Pillow>=8.3.1